# Prevent VirtualJetProducer from wasting time on setup if input empty