From @sethzenz : optimize the VirtualJetProducer for H->gammagamma analysis when multiple vertex hypotheses are used, but many inputs are empty.