@slava77 2015-10-15
@cmsbuild please test