@cmsbuild 2015-10-15
The tests are being triggered in jenkins.
JENKINS_TEST_URL