@cmsbuild 2015-10-15
+1
Tested at: a528c4f7512bbefcb9ba9cfa6bc6c3764a161077
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-11820/8921/summary.html