@cmsbuild 2015-10-16
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-11820/8921/summary.html