@lgray 2015-10-17
@rappoccio @sethzenz Wouldn't it be better to put a single if statement later to skip the expensive stuff instead of three short circuits?