@cvuosalo 2015-10-17
+1

For #11820 a528c4f7512bbefcb9ba9cfa6bc6c3764a161077

Adding skipping of empty inputs to save time in VirtualJetProducer. There should be no change in monitored quantities.

The code changes are satisfactory, and Jenkins [tests](https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-11820/8921/summary.html) against baseline CMSSW_7_6_X_2015-10-15-1100 show no significant differences, as expected. An extended test of workflow 134.702_RunDoubleEG2015B with 70 events against baseline CMSSW_7_6_X_2015-10-14-1600 also shows no significant differences and no significant change in timing, as shown below.
```
Baseline
Max VSIZ 2610.18 on evt 60 ; max RSS 2164.29 on evt 49
Time av 6.69834 s/evt   max 32.8115 s on evt 1
M1 Time av 6.31989 s/evt   max 10.5374 s on evt 47
M8 Time av 6.16307 s/evt   max 10.5374 s on evt 47

PR
Max VSIZ 2610.18 on evt 49 ; max RSS 2199.28 on evt 69
Time av 6.66179 s/evt   max 32.8522 s on evt 1
M1 Time av 6.28222 s/evt   max 10.3564 s on evt 47
M8 Time av 6.10679 s/evt   max 10.3564 s on evt 47
```